Spec Restructuring
This change updates the structure so that it organized into a four major parts.

1. Bindable Services
2. Application Projection
3. Service Request
4. Extensions

There's no significant change to the content of the document as part of this
change, but rather a clear delineation of the different components of the
specification so that iteration can happen clearly on each part independently.
This is a precursor to that discussion and iteration as part of separately
identified changes and inclusion of additional content such as the annotation
examples.

Signed-off-by: Ben Hale <bhale@vmware.com>