Environment Key
Previously, the specification for exposing binding keys as environment
variables used a syntax much like the mapping syntax but without any of the
useful functionality that syntax would have provided.  Specifically, the value
had to be a single mapping from another key in the secret with no additional
modification or substitution.  This made the syntax redundant since the value
could effectively only be a single Secret key name.

This change codifies that behavior by simplifying the syntax and aligning it
with the Kubernetes syntax for [exposing secret keys as environment
variables][k8s].  Note that this change implies that any transformation should
be done as part of `Mapping` and the end result of that transformation should
be exposed directly, effectively making mappings and environment variables
orthogonal to one another while allowing cooperation between the two.

[k8s]: https://kubernetes.io/docs/concepts/configuration/secret/#using-secrets-as-environment-variables

Signed-off-by: Ben Hale <bhale@vmware.com>