Allow the ServiceBinding reflect projected secret on status
Now that mappings are part of the core spec, the ServiceBinding
controller needs to create and manage secrets. It can be useful to
implementors to have a place to stash the name of the projected secret.
To track the secret between reconciliations.

By using the provisioned service duck-type, we are effectively making
the ServiceBinding resource a service for further bindings. While this
may seem odd at first, it enables consolidated mappings to be shared by
several applications.

Signed-off-by: Scott Andrews <andrewssc@vmware.com>