Service Binding Projection Exemplar CRD
This change adds an exemplar CRD for the Service Binding Projection resource.
It does this by adding internal generation support for the type.  This change
also updates the spec a bit so that the voice and section names match the rest
of the document.  There are no semantic changes to the spec.

[resolves #77]

Signed-off-by: Ben Hale <bhale@vmware.com>