Clarify LocalObjectReference usage
The spec previously referred to `.status.binding.name` as a
LocalObjectReference-able. Since the LocalObjectReference object
contains a single field `name`, this would put the resource name as
`.status.binding.name.name`, which is not what our schema and examples
show.

Updating `.status.binding.name` to `.status.binding` since that is the
LocalObjectReference object. This should not be a breaking change since
the intent was clear, even if the language was imprecise.

Signed-off-by: Scott Andrews <andrewssc@vmware.com>