Remove PodSpec-able restriction
The addition of the Application Resource Mapping extension specification
allows us to create a low-effort, cluster-wide mapping of the Service
Binding onto any reasonable (containing env, volumeMount, and volume
concepts) resource type.  Given this new ability, it makes no sense for
the core specification to be restricted to targeting only PodSpec-able
resources.

This change promotes the Application Resource Mapping extension to be
part of the core specification.  It also removes the restrictions in
other parts of the specification that required that application's be
PodSpec-able.

Signed-off-by: Ben Hale <bhale@vmware.com>