Bump CRD version to v1beta1
The next release of the spec is intended as the 1.0 version of the spec,
which will include a commitment to backwards compatibility. The
equivlent version for CRDs is v1beta1.

This is not an indication that the current resoruces as the spec is
written are currently beta quality, but that the next release will be.

Signed-off-by: Scott Andrews <andrewssc@vmware.com>