Restore .spec.type and .spec.provider
While bring back these fields and the general capability, I removed
references that mandated these fields be added to a Secret. Instead, the
requirement is that they are part of the application projection.
Implementors can figure out the best way to make that happen, either by
creating a derivative Secret, or by using a projected volume.

Signed-off-by: Scott Andrews <andrewssc@vmware.com>