Revamp ClusterWorkloadResourceMapping
The mapping structure now more closely aligns with the PodTemplateSpec
shape. It uses a JSONPath query for container-like structures that may
occur anywhere within the target workload resource, and JSON Pointer
references to specific structures either from the root of the object or
relative to a container discovered by a path query.

Signed-off-by: Scott Andrews <andrewssc@vmware.com>