go: mark annotations and volumes as optional
The latest version of the specification (v1) allows for the
`.spec.versions[*].annotations` and `.spec.versions[*].volumes` fields
of `ClusterWorkloadResourceMapping` resources to be optional.  However,
the exemplar go data structures require their existance when
serializing/deserializing.

Update the canonical definition of `ClusterWorkloadResourceMapping`
structs and its generated CRD to be accurate.

Signed-off-by: Andy Sadler <ansadler@redhat.com>