Promote API servicebinding.io/v1beta1 -> v1
The spec and CRDs have been stable for quite some time. We feel
comfortable promoting the v1beta1 APIs to v1. Implementations are free
to continue to support the v1beta1 API for as long as they see fit.

Also updating the scripts to generate the exemplar CRDs to use the
latest dependencies.

Signed-off-by: Scott Andrews <scott@andrews.me>