Clarify binding name collisions
Because service bindings are inherently decentralized, there is no way
to universally detect a name collision. Users are encouraged to avoid
colliding names as the behavior is undefined by the spec.
Implementations may check for known collisions to display a more
meaningful to the user, but detection cannot be guaranteed as it's at
best a good faith attempt.

Signed-off-by: Scott Andrews <scott@andrews.me>